+++ CHANGES.md
@@ -1,9 +1,18 @@
-# v1.1.0
+# v1.2.0
 
 ## New Features
 
 - Add Windows compiler downlaod via github with SHA256
 - Add MacOS and Linux compiler download via github with SHA256 hash
+
+## Bug Fixes
+
+- None yet
+
+# v1.1.0
+
+## New Features
+
 - Add `scripts/sl.cmake` to install sl and a clean compiler
 - Add `scripts/profile.sh` to set the ENV variables from a project
 - add `sos_sdk_copy_file` to copy a file without overwriting